(test) Create Class with Test and Change Tests to fit Testing Convention
