+++ src/test/java/CMVTest.java
@@ -38,36 +38,5 @@ public class CMVTest {
 
   }
 
-  @Test
-  /**
-  * @Test gives True Boolean Value based on LIC5 
-  * It Test that it can find 2 consecitive points has a negative difference
-  **/
-  public void testLIC5True(){
-    Parameters params = new Parameters();
 
-    Point[] dp = new Point[] { new Point(8, 7), new Point(6, 5), new Point(4, 3), new Point(2, 1) };
-
-    CMV cmv = new CMV(params, dp);
-    cmv.populate();
-    assertTrue(cmv.get()[5] == true);
-
-  }
-
-  @Test
-    /**
-  * @Test gives False Boolean Value based on LIC5 
-  * It Test that it can find 2 consecitive points has a negative difference
-  **/
-  public void testLIC5False(){
-
-    Parameters params = new Parameters();
-
-    Point[] dp = new Point[] { new Point(1, 0), new Point(2,0), new Point(3, 0), new Point(4, 0) };
-
-    CMV cmv = new CMV(params, dp);
-    cmv.populate();
-    assertTrue(cmv.get()[5] == false);
-    
-  }
 }
